requirements.txt
Pillow==7.2.0
pyTelegramBotAPI==3.7.2
envparse==0.2.0
